+++ src/lexer/action/action.ts
@@ -5,19 +5,19 @@ import {
   AcceptedActionOutput,
 } from "./model";
 
-export type ActionExec = (input: ActionInput) => ActionOutput;
+export type ActionExec<E> = (input: ActionInput) => ActionOutput<E>;
 
 /**
  * If return a number, the number is how many chars are digested. If the number <= 0, reject.
  */
-export type SimpleActionExec = (
+export type SimpleActionExec<E> = (
   input: ActionInput
-) => number | string | SimpleAcceptedActionOutput;
+) => number | string | SimpleAcceptedActionOutput<E>;
 
-export type ActionSource = RegExp | Action | SimpleActionExec;
+export type ActionSource<E> = RegExp | Action<E> | SimpleActionExec<E>;
 
-export class Action {
-  readonly exec: ActionExec;
+export class Action<E> {
+  readonly exec: ActionExec<E>;
   /**
    * This flag is to indicate whether this action's output might be muted.
    * The lexer will based on this flag to accelerate the lexing process.
@@ -28,17 +28,23 @@ export class Action {
    */
   maybeMuted: boolean;
 
-  constructor(exec: ActionExec, options?: Partial<Pick<Action, "maybeMuted">>) {
+  /**
+   * Don't use this constructor directly, use `Action.from/simple/match` instead.
+   */
+  constructor(
+    exec: ActionExec<E>,
+    options?: Partial<Pick<Action<E>, "maybeMuted">>
+  ) {
     this.exec = exec;
     this.maybeMuted = options?.maybeMuted ?? false;
   }
 
-  static simple(f: SimpleActionExec) {
+  static simple<E>(f: SimpleActionExec<E>): Action<E> {
     return new Action((input) => {
       const res = f(input);
       if (typeof res == "number") {
         if (res <= 0) return { accept: false };
-        return new AcceptedActionOutput({
+        return new AcceptedActionOutput<E>({
           buffer: input.buffer,
           start: input.start,
           muted: false,
@@ -47,7 +53,7 @@ export class Action {
       }
       if (typeof res == "string") {
         if (res.length <= 0) return { accept: false };
-        return new AcceptedActionOutput({
+        return new AcceptedActionOutput<E>({
           buffer: input.buffer,
           start: input.start,
           muted: false,
@@ -58,7 +64,7 @@ export class Action {
       // else, res is SimpleAcceptedActionOutput
       res.digested ??= res.content!.length ?? 0; // if digested is undefined, content must be defined
       if (res.digested <= 0) return { accept: false };
-      return new AcceptedActionOutput({
+      return new AcceptedActionOutput<E>({
         buffer: input.buffer,
         start: input.start,
         muted: res.muted ?? false,
@@ -70,7 +76,7 @@ export class Action {
     });
   }
 
-  static match(
+  static match<E>(
     r: RegExp,
     options?: {
       /**
@@ -84,7 +90,7 @@ export class Action {
        */
       rejectCaret?: boolean;
     }
-  ) {
+  ): Action<E> {
     if (options?.autoSticky ?? true) {
       if (!r.sticky && !r.global)
         // make sure r has the flag 'y/g' so we can use `r.lastIndex` to reset state.
@@ -102,7 +108,7 @@ export class Action {
       r.lastIndex = input.start;
       const res = r.exec(input.buffer);
       if (res && res.index != -1)
-        return new AcceptedActionOutput({
+        return new AcceptedActionOutput<E>({
           muted: false,
           digested: res[0].length,
           buffer: input.buffer,
@@ -113,26 +119,26 @@ export class Action {
     });
   }
 
-  static from(r: ActionSource) {
+  static from<E>(r: ActionSource<E>): Action<E> {
     return r instanceof RegExp
-      ? Action.match(r)
+      ? Action.match<E>(r)
       : r instanceof Action
       ? r
-      : Action.simple(r);
+      : Action.simple<E>(r);
   }
 
   /**
    * Reduce actions to one action.
    * This will reduce the lexer loop times to optimize the performance.
    */
-  static reduce(...actions: ActionSource[]) {
-    return Action.from(actions.reduce((a, b) => Action.from(a).or(b)));
+  static reduce<E>(...actions: ActionSource<E>[]) {
+    return Action.from<E>(actions.reduce((a, b) => Action.from<E>(a).or(b)));
   }
 
   /**
    * Mute action if `accept` is `true` and `muted` is/returned `true`.
    */
-  mute(muted: boolean | ((output: AcceptedActionOutput) => boolean) = true) {
+  mute(muted: boolean | ((output: AcceptedActionOutput<E>) => boolean) = true) {
     if (typeof muted === "boolean")
       return new Action(
         (input) => {
@@ -158,7 +164,7 @@ export class Action {
    * Check the output if `accept` is `true`.
    * `condition` should return error, `undefined` means no error.
    */
-  check(condition: (output: AcceptedActionOutput) => any) {
+  check(condition: (output: AcceptedActionOutput<E>) => E) {
     return new Action((buffer) => {
       const output = this.exec(buffer);
       if (output.accept)
@@ -181,7 +187,9 @@ export class Action {
   /**
    * Reject if `accept` is `true` and `rejecter` is/returns `true`.
    */
-  reject(rejecter: boolean | ((output: AcceptedActionOutput) => any) = true) {
+  reject(
+    rejecter: boolean | ((output: AcceptedActionOutput<E>) => any) = true
+  ) {
     if (typeof rejecter === "boolean") {
       if (rejecter) return new Action(() => ({ accept: false })); // always reject
       return this; // just return self
@@ -199,7 +207,7 @@ export class Action {
   /**
    * Call `f` if `accept` is `true`.
    */
-  then(f: (output: AcceptedActionOutput) => void) {
+  then(f: (output: AcceptedActionOutput<E>) => void) {
     return new Action((buffer) => {
       const output = this.exec(buffer);
       if (output.accept) f(output);
@@ -211,7 +219,7 @@ export class Action {
    * Execute the new action if current action can't accept input.
    * Sadly there is no operator overloading in typescript.
    */
-  or(a: ActionSource) {
+  or(a: ActionSource<E>) {
     const other = Action.from(a);
     return new Action(
       (buffer) => {