+++ src/parser/ELR/model/re-lex.ts
@@ -5,7 +5,7 @@ import { Callback, ParserContext } from "./context";
 export type ReLexStack<State, T> = {
   readonly stateStack: State[];
   readonly buffer: ASTNode<T>[];
-  readonly lexer: ILexer;
+  readonly lexer: ILexer<any>;
   readonly index: number;
   readonly errors: ASTNode<T>[];
   readonly rollbackStackLength: number;