+++ tests/lexer/lexer.test.ts
@@ -52,7 +52,7 @@ test("number", () => {
       content: str,
       start: 0,
       error: undefined,
-    } as Token);
+    } as Token<any>);
   });
 });
 
@@ -69,7 +69,7 @@ test("anonymous", () => {
       content: str,
       start: 0,
       error: undefined,
-    } as Token);
+    } as Token<any>);
   });
 });